add Caffeine for cache management, openapi for documentation, problemDetails for exception handling, minor refactors

+++ src/main/java/com/ayagmar/jobapplicationtracker/service/CityService.java
@@ -11,6 +11,9 @@ import com.ayagmar.jobapplicationtracker.repository.CityRepository;
 import com.ayagmar.jobapplicationtracker.repository.CountryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.cache.annotation.CacheConfig;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -19,6 +22,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Slf4j
 @RequiredArgsConstructor
+@CacheConfig(cacheNames = "cities")
 public class CityService {
     private final CityRepository cityRepository;
     private final CountryRepository countryRepository;
@@ -37,6 +41,7 @@ public class CityService {
     }
 
     @Transactional(readOnly = true)
+    @Cacheable
     public PaginatedResponse<CityResponse> getAllCities(Pageable pageable) {
         Page<City> cities = cityRepository.findAll(pageable);
         log.info("Retrieved {} cities", cities.getTotalElements());
@@ -45,6 +50,7 @@ public class CityService {
     }
 
     @Transactional(readOnly = true)
+    @Cacheable(key = "#id")
     public CityResponse getCityById(Long id) {
         var city = cityRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("city " + id + " is not found"));
@@ -52,11 +58,16 @@ public class CityService {
     }
 
     @Transactional
+    @CacheEvict(key = "#id", value = "cities")
     public void deleteCity(Long id) {
-        getCityById(id);
-        cityRepository.deleteById(id);
-        log.info("City with id " + id + " is deleted");
+        if (cityRepository.existsById(id)) {
+            cityRepository.deleteById(id);
+            log.info("City with id {} is deleted", id);
+        } else {
+            log.error("City with id {} does not exist", id);
+        }
     }
+
 
     private void validateName(String name) {
         cityRepository.findCityByName(name).ifPresent(company -> {
